chore: script to lint mdx files by parsing it
To implement auto translation, we need to create a script to
separate some nodes such as import statement and code block using
`remark` package. For that script to work, the mdx files need
to be parseable by `remark`. Therefore, we need to check before commit
whether all mdx files are parseable.