change prettier printWidth for .md and .mdx files.
The whitespace `{" "}` JavaScript expression will be treated as
text by Deepl. Sometimes, it will be relocated (e.g. in Japanese)
from the end of line to the middle of line which is not desired.

By setting the printWidth to Infinity, the text will not be wrapped
to next line, so there is no need for the whitespace `{" "}` expression.