remove package.json change as translation action trigger
At first, it was added to handle scenarios when docusaurus and deepmark
version are aupgraded. Because sometimes docusaurus and deepmark
version change also bring generated translation changes.

But, the downside is it will also be triggered when other deps are
upgraded. So, it's better to manually run the action from the github
action dashboard when such docusarus and deepmark version change is
happening.